extract procedure to create my channel
